refactor(runtime): extract max-iteration exit into agent/turn — #7415
The post-loop tail moves to
turn/max_iter.rs::finish_after_max_iterations: exhaustion logging,
summary-request user message, tools-free chat call with
step-timeout/cancel select, and the Ok(accumulated + summary) or
exceeded-maximum-tool-iterations bail. loop_.rs drops its now-unused
Duration import (tests import it separately).
